api/requirements.txt
-i https://pypi.org/simple
asgiref==3.11.1; python_version >= '3.9'
django==6.0.3; python_version >= '3.12'
django-cors-headers==4.9.0; python_version >= '3.9'
django-debug-toolbar==6.2.0; python_version >= '3.10'
django-filter==25.2; python_version >= '3.10'
djangorestframework==3.17.0; python_version >= '3.10'
drf-nested-routers==0.95.0; python_version >= '3.8'
psycopg[binary]==3.3.3; python_version >= '3.10'
psycopg-binary==3.3.3
sqlparse==0.5.5; python_version >= '3.8'
typing-extensions==4.15.0; python_version < '3.13'
